streamlit-app/requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
requests==2.31.0
plotly==5.18.0
